refactor(be): Phase 1 - Extract JniDataBridge utility class from JniConnector
Extract all static data exchange methods from JniConnector into a new
JniDataBridge utility class. JniDataBridge is a pure stateless utility
for C++ Block <-> JNI shared memory data exchange, data-source agnostic.

Moved to JniDataBridge:
- TableMetaAddress helper class
- fill_block / fill_column (Java -> C++ Block)
- to_java_table / parse_table_schema (C++ Block -> Java)
- get_jni_type / get_jni_type_with_different_string (type mapping)
- All private _fill_*_column helpers and templates

JniConnector retains backward-compatible inline wrappers that delegate
to JniDataBridge, so all existing callers (vjdbc_connector,
vjni_format_transformer, UDF functions, 6 JNI readers) continue to
work without changes.

This is Phase 1 of the JNI read/write architecture refactoring plan.
